refactor(form-builder): colocate inline editing hook in tag options control
Move useInlineEditableOptionRows into JsonFormsTagCategoryOptionsControl
since it has a single caller, keeping the PR diff easier to review.
